backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
sqlalchemy==2.0.23
python-multipart==0.0.6
python-dotenv==1.0.0
email-validator==2.1.0
aiosmtplib==3.0.1
dateparser==1.2.0
httpx==0.25.2
requests==2.31.0
# AI Integration
openai==1.3.0
# vonage removed - SMS functionality disabled
pytest==7.4.3
pytest-asyncio==0.21.1